network-generator/pyproject.toml
[project]
name = "placement-explain-datagen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "clingexplaid>=1.1.0",
    "clingo>=5.7.1",
    "networkx>=3.4.2",
    "numpy>=2.2.3",
    "pyyaml>=6.0.2",
]

[tool.setuptools]
packages = ["random_network_generator"]

[tool.setuptools.package-data]
"random_network_generator" = ["*"]
